src/requirements.txt
# Importante: pypdf é o sucessor de PyPDF2 e PyPDF3, atualmente mantido
pypdf>=3.15.1
pdfminer.six>=20221105
requests>=2.31.0
beautifulsoup4>=4.12.2
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0
PyYAML>=6.0
anthropic>=0.7.0
openai>=1.1.1